requirements.txt
Django==4.2.2
requests==2.29
types-requests==2.27.16
drf_yasg==1.21.6
django_debug_toolbar==3.2.4
djangorestframework==3.14.0
django-split-settings==1.1.0
django-environ==0.8.1
django-filter==21.1
django-simple-history==3.0.0
djangorestframework-simplejwt==5.1.0
psycopg[binary]==3.1.9
gunicorn==20.1.0
django-summernote==0.8.20.0
python-dotenv==1.0.0
